Edit view livewire crm customer

+++ resources/views/livewire/crm/customer/customer-lists.blade.php
@@ -23,14 +23,19 @@
                     </div>
                     <div class="col-3 d-flex justify-content-center">
                         <div class="btn-group w-100" role="group">
-                            <button wire:click="$dispatch('refresh-customer-ax')" type="button"
-                                class="btn btn-secondary" data-toggle="modal" data-target="#modal-customer-ax">
-                                <i class="fas fa-plus"></i> AX
-                            </button>
-                            <button type="button" class="btn btn-info" data-toggle="modal"
-                                data-target="#modal-add-customer">
-                                <i class="fas fa-plus"></i> Add
-                            </button>
+
+                            @can('customer.create')
+                                <button wire:click="$dispatch('refresh-customer-ax')" type="button"
+                                    class="btn btn-secondary" data-toggle="modal" data-target="#modal-customer-ax">
+                                    <i class="fas fa-plus"></i> AX
+                                </button>
+
+                                <button type="button" class="btn btn-info" data-toggle="modal"
+                                    data-target="#modal-add-customer">
+                                    <i class="fas fa-plus"></i> Add
+                                </button>
+                            @endcan
+
                             <button wire:click="$dispatch('refresh-customer')" type="button" class="btn btn-success">
                                 <i class="fas fa-sync-alt"></i>
                             </button>
@@ -54,8 +59,8 @@
                 </div>
 
                 <div class="table-responsive">
-                    <table class="table table-hover table-sm">
-                        <thead>
+                    <table class="table table-sm table-hover table-bordered">
+                        <thead class="table-info">
                             <th scope="col">#</th>
                             <th scope="col">Created</th>
                             <th scope="col">Updated</th>
@@ -65,7 +70,13 @@
                             <th scope="col">Name English</th>
                             <th scope="col">Name Thai</th>
                             <th scope="col">Parent</th>
-                            <th scope="col" style="width: 115px">Action</th>
+                            @can('crm.create')
+                                <th scope="col" colspan="3" class="text-center">Action</th>
+                            @elsecan('customer.edit')
+                                <th scope="col" colspan="3" class="text-center">Action</th>
+                            @elsecan('customer.delete')
+                                <th scope="col" colspan="3" class="text-center">Action</th>
+                            @endcan
                         </thead>
 
                         <tbody>
@@ -108,7 +119,7 @@
                                         </div>
 
                                     </td>
-                                    <td>
+                                    <td class="text-center">
                                         @if ($item->source === '0')
                                             <span class="badge badge-pill badge-primary">
                                                 AX
@@ -128,38 +139,68 @@
                                     <td>{{ $item->name_english }}</td>
                                     <td>{{ $item->name_thai }}</td>
                                     <td>{{ $item->parent_code }}</td>
-                                    <td>
-                                        <button
-                                            wire:click.prevent="$dispatch('select-customer',
+
+                                    @if ($item->source === '1' || $item->source === '2')
+                                        @can('crm.create')
+                                            <td style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="$dispatch('select-customer',{id:{{ $item->id }}})"
+                                                    class="btn btn-success btn-sm">
+                                                    <i class="fas fa-check"></i>
+                                                </button>
+                                            </td>
+                                        @endcan
+
+                                        @can('customer.edit')
+                                            <td style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="$dispatch('edit-customer',{id:{{ $item->id }}})"
+                                                    type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
+                                                    data-target="#modal-edit-customer">
+                                                    <i class="fas fa-edit"></i>
+                                                </button>
+                                            </td>
+                                        @endcan
+
+                                        @can('customer.delete')
+                                            <td style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="deleteCustomer({{ $item->id }},{{ "'" . str_replace("'", '', $item->name_english) . "'" }})"
+                                                    class="btn btn-sm btn-danger">
+                                                    <i class="fas fa-trash"></i>
+                                                </button>
+                                            </td>
+                                        @endcan
+                                    @else
+                                        {{-- Data Source = 0 AX --}}
+                                        @can('crm.create')
+                                            <td colspan="3" style="width: 40px" class="p-1 text-center">
+                                                <button
+                                                    wire:click.prevent="$dispatch('select-customer',
                                             {id:{{ $item->id }}})"
-                                            wire:click="$dispatch('refresh-customer')" class="btn btn-success btn-sm">
-                                            <i class="fas fa-check"></i>
-                                        </button>
-
-                                        @if ($item->source === '1' || $item->source === '2')
-                                            <button
-                                                wire:click.prevent="$dispatch('edit-customer',{id:{{ $item->id }}})"
-                                                type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
-                                                data-target="#modal-edit-customer">
-                                                <i class="fas fa-edit"></i>
-                                            </button>
-                                            <button
-                                                wire:click.prevent="deleteCustomer({{ $item->id }},{{ "'" . str_replace("'", '', $item->name_english) . "'" }})"
-                                                {{-- wire:click.prevent="deleteCustomer({{ $item->id }},{{ "'" . $item->name_english . "'" }})" --}} class="btn btn-sm btn-danger">
-                                                <i class="fas fa-trash"></i>
-                                            </button>
-                                        @endif
-
-                                    </td>
+                                                    wire:click="$dispatch('refresh-customer')"
+                                                    class="btn btn-success btn-sm">
+                                                    <i class="fas fa-check"></i>
+                                                </button>
+                                            </td>
+                                        @elsecan('customer.edit')
+                                            <td colspan="3"></td>
+                                        @elsecan('customer.delete')
+                                            <td></td>
+                                        @endcan
+                                    @endif
                                 </tr>
                             @endforeach
 
                             @livewire('crm.customer.customer-edit')
 
-                            {{-- @livewire('crm.customer-edit') --}}
-
                         </tbody>
                     </table>
+
+                    @livewire('crm.customer.customer-ax-lists')
+
+                    @livewire('crm.customer.customer-create')
+
                 </div>
 
             </div>
@@ -179,9 +220,6 @@
 @script
     <script>
         $wire.on("confirm-delete-customer", (event) => {
-
-            // alert(event.name);
-
             Swal.fire({
                 title: "Are you sure delete ?",
                 text: `Customer : ${event.name}`,
@@ -205,7 +243,6 @@
     <script>
         document.addEventListener('livewire:initialized', () => {
             @this.on('close-modal-customer', (event) => {
-                // alert('Close Modal')
                 setTimeout(() => {
                     $('#modal-edit-customer').modal('hide')
                 }, 3000);
